+++ src/main/java/org/eclipse/keyple/card/calypso/CmdCardOpenSession.java
@@ -16,6 +16,7 @@ import static org.eclipse.keyple.card.calypso.DtoAdapters.*;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
+import org.calypsonet.terminal.calypso.WriteAccessLevel;
 import org.calypsonet.terminal.card.ApduResponseApi;
 import org.eclipse.keyple.core.util.ApduUtil;
 import org.eclipse.keyple.core.util.ByteArrayUtil;
@@ -30,7 +31,7 @@ import org.slf4j.LoggerFactory;
 final class CmdCardOpenSession extends CardCommand {
 
   private static final Logger logger = LoggerFactory.getLogger(CmdCardOpenSession.class);
-  private static final String EXTRA_INFO_FORMAT = "KEYINDEX:%d, SFI:%02Xh, REC:%d";
+  private static final String EXTRA_INFO_FORMAT = "KEYINDEX:%d, SFI:%02Xh, REC:%d, PREOPEN:%s";
 
   private static final Map<Integer, StatusProperties> STATUS_TABLE;
 
@@ -71,35 +72,43 @@ final class CmdCardOpenSession extends CardCommand {
     m.put(
         0x6B00,
         new StatusProperties(
-            "P1 or P2 value not supported (key index incorrect, wrong P2).",
+            "P1 or P2 value not supported (key index incorrect, wrong P2, extended mode not supported).",
             CardIllegalParameterException.class));
-    m.put(0x61FF, new StatusProperties("Correct execution (ISO7816 T=0).", null));
+    m.put(0x61FF, new StatusProperties("Correct execution (ISO7816 T=0)."));
+    m.put(
+        0x6200,
+        new StatusProperties(
+            "Successful execution, with warning (Pre-Open variant, secure session not opened)."));
     STATUS_TABLE = m;
   }
 
   private final boolean isExtendedModeAllowed;
+  private final boolean isPreOpenMode;
+  private final WriteAccessLevel writeAccessLevel;
+  private final int sfi;
+  private final int recordNumber;
 
-  private int sfi;
-  private int recordNumber;
-
-  /** The secure session. */
-  private SecureSession secureSession;
+  private boolean isPreviousSessionRatified;
+  private byte[] challengeTransactionCounter;
+  private byte[] challengeRandomNumber;
+  private Byte kif;
+  private Byte kvc;
+  private byte[] recordData;
 
   /**
    * Constructor.
    *
    * @param calypsoCard The Calypso card.
-   * @param keyIndex The key index.
+   * @param writeAccessLevel The write access level.
    * @param samChallenge The SAM challenge.
    * @param sfi The optional SFI of the file to read.
    * @param recordNumber The optional record number to read.
    * @param isExtendedModeAllowed True if the extended mode is allowed.
-   * @throws IllegalArgumentException If the key index is 0 and rev is 2.4
    * @since 2.0.1
    */
   CmdCardOpenSession(
       CalypsoCardAdapter calypsoCard,
-      byte keyIndex,
+      WriteAccessLevel writeAccessLevel,
       byte[] samChallenge,
       int sfi,
       int recordNumber,
@@ -108,17 +117,23 @@ final class CmdCardOpenSession extends CardCommand {
     super(CardCommandRef.OPEN_SECURE_SESSION, 0, calypsoCard);
 
     this.isExtendedModeAllowed = isExtendedModeAllowed;
+    this.isPreOpenMode = false;
+    this.writeAccessLevel = writeAccessLevel;
+    this.sfi = sfi;
+    this.recordNumber = recordNumber;
+
+    byte keyIndex = (byte) (writeAccessLevel.ordinal() + 1);
     switch (getCalypsoCard().getProductType()) {
       case PRIME_REVISION_1:
-        createRev10(keyIndex, samChallenge, sfi, recordNumber);
+        createRev10(keyIndex, samChallenge);
         break;
       case PRIME_REVISION_2:
-        createRev24(keyIndex, samChallenge, sfi, recordNumber);
+        createRev24(keyIndex, samChallenge);
         break;
       case PRIME_REVISION_3:
       case LIGHT:
       case BASIC:
-        createRev3(keyIndex, samChallenge, sfi, recordNumber);
+        createRev3(keyIndex, samChallenge);
         break;
       default:
         throw new IllegalArgumentException(
@@ -127,18 +142,38 @@ final class CmdCardOpenSession extends CardCommand {
   }
 
   /**
+   * Constructor.
+   *
+   * @param writeAccessLevel The write access level.
+   * @param sfi The optional SFI of the file to read.
+   * @param recordNumber The optional record number to read.
+   * @param isExtendedModeAllowed True if the extended mode is allowed.
+   * @throws IllegalArgumentException If the key index is 0 and rev is 2.4
+   * @since 2.3.2
+   */
+  CmdCardOpenSession(
+      WriteAccessLevel writeAccessLevel, int sfi, int recordNumber, boolean isExtendedModeAllowed) {
+
+    super(CardCommandRef.OPEN_SECURE_SESSION, 0, null);
+
+    this.isExtendedModeAllowed = isExtendedModeAllowed;
+    this.isPreOpenMode = true;
+    this.writeAccessLevel = writeAccessLevel;
+    this.sfi = sfi;
+    this.recordNumber = recordNumber;
+
+    byte keyIndex = (byte) (writeAccessLevel.ordinal() + 1);
+    createRev3(keyIndex, new byte[0]);
+  }
+
+  /**
    * Create Rev 3
    *
    * @param keyIndex the key index.
    * @param samChallenge the sam challenge returned by the SAM Get Challenge APDU command.
-   * @param sfi the sfi to select.
-   * @param recordNumber the record number to read.
    * @throws IllegalArgumentException If the request is inconsistent
    */
-  private void createRev3(byte keyIndex, byte[] samChallenge, int sfi, int recordNumber) {
-
-    this.sfi = sfi;
-    this.recordNumber = recordNumber;
+  private void createRev3(byte keyIndex, byte[] samChallenge) {
 
     byte p1 = (byte) ((recordNumber * 8) + keyIndex);
     byte p2;
@@ -168,7 +203,8 @@ final class CmdCardOpenSession extends CardCommand {
                 (byte) 0)));
 
     if (logger.isDebugEnabled()) {
-      String extraInfo = String.format(EXTRA_INFO_FORMAT, keyIndex, sfi, recordNumber);
+      String extraInfo =
+          String.format(EXTRA_INFO_FORMAT, keyIndex, sfi, recordNumber, isPreOpenMode);
       addSubName(extraInfo);
     }
   }
@@ -178,22 +214,9 @@ final class CmdCardOpenSession extends CardCommand {
    *
    * @param keyIndex the key index.
    * @param samChallenge the sam challenge returned by the SAM Get Challenge APDU command.
-   * @param sfi the sfi to select.
-   * @param recordNumber the record number to read.
-   * @throws IllegalArgumentException If key index is 0 (rev 2.4)
-   * @throws IllegalArgumentException If the request is inconsistent
    */
-  private void createRev24(byte keyIndex, byte[] samChallenge, int sfi, int recordNumber) {
-
-    if (keyIndex == 0x00) {
-      throw new IllegalArgumentException("Key index can't be zero for rev 2.4!");
-    }
-
-    this.sfi = sfi;
-    this.recordNumber = recordNumber;
-
+  private void createRev24(byte keyIndex, byte[] samChallenge) {
     byte p1 = (byte) (0x80 + (recordNumber * 8) + keyIndex);
-
     buildLegacyApduRequest(keyIndex, samChallenge, sfi, recordNumber, p1);
   }
 
@@ -202,22 +225,9 @@ final class CmdCardOpenSession extends CardCommand {
    *
    * @param keyIndex the key index.
    * @param samChallenge the sam challenge returned by the SAM Get Challenge APDU command.
-   * @param sfi the sfi to select.
-   * @param recordNumber the record number to read.
-   * @throws IllegalArgumentException If key index is 0 (rev 1.0)
-   * @throws IllegalArgumentException If the request is inconsistent
    */
-  private void createRev10(byte keyIndex, byte[] samChallenge, int sfi, int recordNumber) {
-
-    if (keyIndex == 0x00) {
-      throw new IllegalArgumentException("Key index can't be zero for rev 1.0!");
-    }
-
-    this.sfi = sfi;
-    this.recordNumber = recordNumber;
-
+  private void createRev10(byte keyIndex, byte[] samChallenge) {
     byte p1 = (byte) ((recordNumber * 8) + keyIndex);
-
     buildLegacyApduRequest(keyIndex, samChallenge, sfi, recordNumber, p1);
   }
 
@@ -250,7 +260,8 @@ final class CmdCardOpenSession extends CardCommand {
                 (byte) 0)));
 
     if (logger.isDebugEnabled()) {
-      String extraInfo = String.format(EXTRA_INFO_FORMAT, keyIndex, sfi, recordNumber);
+      String extraInfo =
+          String.format(EXTRA_INFO_FORMAT, keyIndex, sfi, recordNumber, isPreOpenMode);
       addSubName(extraInfo);
     }
   }
@@ -267,28 +278,20 @@ final class CmdCardOpenSession extends CardCommand {
   }
 
   /**
-   * @return the SFI of the file read while opening the secure session
-   * @since 2.0.1
-   */
-  int getSfi() {
-    return sfi;
-  }
-
-  /**
-   * @return the record number to read
-   * @since 2.0.1
-   */
-  int getRecordNumber() {
-    return recordNumber;
-  }
-
-  /**
    * {@inheritDoc}
    *
    * @since 2.0.1
    */
   @Override
   void parseApduResponse(ApduResponseApi apduResponse) throws CardCommandException {
+    if (isPreOpenMode) {
+      getCalypsoCard()
+          .setPreOpenWriteAccessLevel(writeAccessLevel)
+          .setPreOpenUseExtendedMode(isExtendedModeAllowed)
+          .setPreOpenSfi((byte) sfi)
+          .setPreOpenRecordNumber(recordNumber)
+          .setPreOpenDataOut(apduResponse.getDataOut());
+    }
     super.parseApduResponse(apduResponse);
     byte[] dataOut = getApduResponse().getDataOut();
     switch (getCalypsoCard().getProductType()) {
@@ -302,13 +305,12 @@ final class CmdCardOpenSession extends CardCommand {
         parseRev3(dataOut);
     }
     // CL-CSS-INFORAT.1
-    getCalypsoCard().setDfRatified(secureSession.isPreviousSessionRatified());
+    getCalypsoCard().setDfRatified(isPreviousSessionRatified);
     // CL-CSS-INFOTCNT.1
     getCalypsoCard()
-        .setTransactionCounter(
-            ByteArrayUtil.extractInt(secureSession.getChallengeTransactionCounter(), 0, 3, false));
-    if (secureSession.getOriginalData().length > 0) {
-      getCalypsoCard().setContent((byte) sfi, recordNumber, secureSession.getOriginalData());
+        .setTransactionCounter(ByteArrayUtil.extractInt(challengeTransactionCounter, 0, 3, false));
+    if (recordData.length > 0) {
+      getCalypsoCard().setContent((byte) sfi, recordNumber, recordData);
     }
   }
 
@@ -318,41 +320,26 @@ final class CmdCardOpenSession extends CardCommand {
    * @param apduResponseData The response data.
    */
   private void parseRev3(byte[] apduResponseData) {
-
-    boolean previousSessionRatified;
-    boolean manageSecureSessionAuthorized;
     int offset;
-
     // CL-CSS-OSSRFU.1
     if (isExtendedModeAllowed) {
       offset = 4;
-      previousSessionRatified = (apduResponseData[8] & 0x01) == (byte) 0x00;
-      manageSecureSessionAuthorized = (apduResponseData[8] & 0x02) == (byte) 0x02;
+      isPreviousSessionRatified = (apduResponseData[8] & 0x01) == (byte) 0x00;
+      boolean manageSecureSessionAuthorized = (apduResponseData[8] & 0x02) == (byte) 0x02;
       if (!manageSecureSessionAuthorized) {
         getCalypsoCard().disableExtendedMode();
       }
     } else {
       offset = 0;
-      previousSessionRatified = (apduResponseData[4] == (byte) 0x00);
-      manageSecureSessionAuthorized = false;
+      isPreviousSessionRatified = (apduResponseData[4] == (byte) 0x00);
       getCalypsoCard().disableExtendedMode();
     }
-
-    byte kif = apduResponseData[5 + offset];
-    byte kvc = apduResponseData[6 + offset];
+    challengeTransactionCounter = Arrays.copyOfRange(apduResponseData, 0, 3);
+    challengeRandomNumber = Arrays.copyOfRange(apduResponseData, 3, 4 + offset);
+    kif = apduResponseData[5 + offset];
+    kvc = apduResponseData[6 + offset];
     int dataLength = apduResponseData[7 + offset];
-    byte[] data = Arrays.copyOfRange(apduResponseData, 8 + offset, 8 + offset + dataLength);
-
-    this.secureSession =
-        new SecureSession(
-            Arrays.copyOfRange(apduResponseData, 0, 3),
-            Arrays.copyOfRange(apduResponseData, 3, 4 + offset),
-            previousSessionRatified,
-            manageSecureSessionAuthorized,
-            kif,
-            kvc,
-            data,
-            apduResponseData);
+    recordData = Arrays.copyOfRange(apduResponseData, 8 + offset, 8 + offset + dataLength);
   }
 
   /**
@@ -383,45 +370,31 @@ final class CmdCardOpenSession extends CardCommand {
    * @param apduResponseData The response data.
    */
   private void parseRev24(byte[] apduResponseData) {
-
-    boolean previousSessionRatified;
-
-    byte[] data;
-
     switch (apduResponseData.length) {
       case 5:
-        previousSessionRatified = true;
-        data = new byte[0];
+        isPreviousSessionRatified = true;
+        recordData = new byte[0];
         break;
       case 34:
-        previousSessionRatified = true;
-        data = Arrays.copyOfRange(apduResponseData, 5, 34);
+        isPreviousSessionRatified = true;
+        recordData = Arrays.copyOfRange(apduResponseData, 5, 34);
         break;
       case 7:
-        previousSessionRatified = false;
-        data = new byte[0];
+        isPreviousSessionRatified = false;
+        recordData = new byte[0];
         break;
       case 36:
-        previousSessionRatified = false;
-        data = Arrays.copyOfRange(apduResponseData, 7, 36);
+        isPreviousSessionRatified = false;
+        recordData = Arrays.copyOfRange(apduResponseData, 7, 36);
         break;
       default:
         throw new IllegalStateException(
             "Bad response length to Open Secure Session: " + apduResponseData.length);
     }
-
-    byte kvc = apduResponseData[0];
-
-    this.secureSession =
-        new SecureSession(
-            Arrays.copyOfRange(apduResponseData, 1, 4),
-            Arrays.copyOfRange(apduResponseData, 4, 5),
-            previousSessionRatified,
-            false,
-            null,
-            kvc,
-            data,
-            apduResponseData);
+    challengeTransactionCounter = Arrays.copyOfRange(apduResponseData, 1, 4);
+    challengeRandomNumber = Arrays.copyOfRange(apduResponseData, 4, 5);
+    kif = null;
+    kvc = apduResponseData[0];
   }
 
   /**
@@ -451,44 +424,31 @@ final class CmdCardOpenSession extends CardCommand {
    * @param apduResponseData The response data.
    */
   private void parseRev10(byte[] apduResponseData) {
-
-    boolean previousSessionRatified;
-
-    byte[] data;
-
     switch (apduResponseData.length) {
       case 4:
-        previousSessionRatified = true;
-        data = new byte[0];
+        isPreviousSessionRatified = true;
+        recordData = new byte[0];
         break;
       case 33:
-        previousSessionRatified = true;
-        data = Arrays.copyOfRange(apduResponseData, 4, 33);
+        isPreviousSessionRatified = true;
+        recordData = Arrays.copyOfRange(apduResponseData, 4, 33);
         break;
       case 6:
-        previousSessionRatified = false;
-        data = new byte[0];
+        isPreviousSessionRatified = false;
+        recordData = new byte[0];
         break;
       case 35:
-        previousSessionRatified = false;
-        data = Arrays.copyOfRange(apduResponseData, 6, 35);
+        isPreviousSessionRatified = false;
+        recordData = Arrays.copyOfRange(apduResponseData, 6, 35);
         break;
       default:
         throw new IllegalStateException(
             "Bad response length to Open Secure Session: " + apduResponseData.length);
     }
-
-    /* KVC doesn't exist and is set to null for this type of card */
-    this.secureSession =
-        new SecureSession(
-            Arrays.copyOfRange(apduResponseData, 0, 3),
-            Arrays.copyOfRange(apduResponseData, 3, 4),
-            previousSessionRatified,
-            false,
-            null,
-            null,
-            data,
-            apduResponseData);
+    challengeTransactionCounter = Arrays.copyOfRange(apduResponseData, 0, 3);
+    challengeRandomNumber = Arrays.copyOfRange(apduResponseData, 3, 4);
+    kif = null;
+    kvc = null;
   }
 
   /**
@@ -496,31 +456,23 @@ final class CmdCardOpenSession extends CardCommand {
    * @since 2.0.1
    */
   byte[] getCardChallenge() {
-    return secureSession.getChallengeRandomNumber();
-  }
-
-  /**
-   * @return True if the managed secure session is authorized.
-   * @since 2.0.1
-   */
-  boolean isManageSecureSessionAuthorized() {
-    return secureSession.isManageSecureSessionAuthorized();
+    return challengeRandomNumber;
   }
 
   /**
    * @return The current KIF.
    * @since 2.0.1
    */
-  Byte getSelectedKif() {
-    return secureSession.getKIF();
+  Byte getKif() {
+    return kif;
   }
 
   /**
    * @return The current KVC.
    * @since 2.0.1
    */
-  Byte getSelectedKvc() {
-    return secureSession.getKVC();
+  Byte getKvc() {
+    return kvc;
   }
 
   /**
@@ -531,135 +483,5 @@ final class CmdCardOpenSession extends CardCommand {
   @Override
   Map<Integer, StatusProperties> getStatusTable() {
     return STATUS_TABLE;
-  }
-
-  /** The Class SecureSession. */
-  private static class SecureSession {
-
-    /** Challenge transaction counter */
-    private final byte[] challengeTransactionCounter;
-
-    /** Challenge random number */
-    private final byte[] challengeRandomNumber;
-
-    /** The previous session ratified boolean. */
-    private final boolean previousSessionRatified;
-
-    /** The manage secure session authorized boolean. */
-    private final boolean manageSecureSessionAuthorized;
-
-    /** The kif (it may be null if it doesn't exist in the considered card [rev 1.0]). */
-    private final Byte kif;
-
-    /** The kvc (it may be null if it doesn't exist in the considered card [rev 1.0]). */
-    private final Byte kvc;
-
-    /** The original data. */
-    private final byte[] originalData;
-
-    /** The secure session data. */
-    private final byte[] secureSessionData;
-
-    /**
-     * Instantiates a new SecureSession
-     *
-     * @param challengeTransactionCounter Challenge transaction counter.
-     * @param challengeRandomNumber Challenge random number.
-     * @param previousSessionRatified the previous session ratified.
-     * @param manageSecureSessionAuthorized the manage secure session authorized.
-     * @param kif the KIF from the response of the open secure session APDU command.
-     * @param kvc the KVC from the response of the open secure session APDU command.
-     * @param originalData the original data from the response of the open secure session APDU.
-     *     command
-     * @param secureSessionData the secure session data from the response of open secure session.
-     *     APDU command
-     * @since 2.0.1
-     */
-    private SecureSession( // NOSONAR
-        byte[] challengeTransactionCounter,
-        byte[] challengeRandomNumber,
-        boolean previousSessionRatified,
-        boolean manageSecureSessionAuthorized,
-        Byte kif,
-        Byte kvc,
-        byte[] originalData,
-        byte[] secureSessionData) {
-      this.challengeTransactionCounter = challengeTransactionCounter;
-      this.challengeRandomNumber = challengeRandomNumber;
-      this.previousSessionRatified = previousSessionRatified;
-      this.manageSecureSessionAuthorized = manageSecureSessionAuthorized;
-      this.kif = kif;
-      this.kvc = kvc;
-      this.originalData = originalData;
-      this.secureSessionData = secureSessionData;
-    }
-
-    public byte[] getChallengeTransactionCounter() {
-      return challengeTransactionCounter;
-    }
-
-    public byte[] getChallengeRandomNumber() {
-      return challengeRandomNumber;
-    }
-
-    /**
-     * Checks if is previous session ratified.
-     *
-     * @return The boolean
-     * @since 2.0.1
-     */
-    public boolean isPreviousSessionRatified() {
-      return previousSessionRatified;
-    }
-
-    /**
-     * Checks if is manage secure session authorized.
-     *
-     * @return True if the secure session is authorized
-     * @since 2.0.1
-     */
-    public boolean isManageSecureSessionAuthorized() {
-      return manageSecureSessionAuthorized;
-    }
-
-    /**
-     * Gets the kif.
-     *
-     * @return A byte
-     * @since 2.0.1
-     */
-    public Byte getKIF() {
-      return kif;
-    }
-
-    /**
-     * Gets the kvc.
-     *
-     * @return A byte
-     * @since 2.0.1
-     */
-    public Byte getKVC() {
-      return kvc;
-    }
-
-    /**
-     * Gets the original data.
-     *
-     * @return An array of bytes
-     * @since 2.0.1
-     */
-    public byte[] getOriginalData() {
-      return originalData;
-    }
-
-    /**
-     * Gets the secure session data.
-     *
-     * @return An array of bytes
-     * @since 2.0.1
-     */
-    public byte[] getSecureSessionData() {
-      return secureSessionData;
-    }
   }
 }