+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- The errors raised by the "Read Record" command executed during the card selection process are no longer blocking.
 
 ## [2.3.1] - 2022-12-22
 ### Fixed